Updated mod options - Moved class `CLOptions.OptionBuilder` to `Utils.OptionBuilder` - Added new option: Invert Controls
